revision for new plugin
