+++ neural_chat/chatbot.py
@@ -31,11 +31,9 @@ from .pipeline.plugins.audio.asr import AudioSpeechRecognition
 from .pipeline.plugins.audio.asr_chinese import ChineseAudioSpeechRecognition
 from .pipeline.plugins.audio.tts import TextToSpeech
 from .pipeline.plugins.audio.tts_chinese import ChineseTextToSpeech
-from .pipeline.plugins.retrievers.indexing.document_parser import DocumentIndexing
-from .pipeline.plugins.retrievers.retriever.langchain import ChromaRetriever
-from .pipeline.plugins.retrievers.retriever import BM25Retriever
 from .pipeline.plugins.security.safety_checker import SafetyChecker
 from .pipeline.plugins.intent_detector import IntentDetector
+from .pipeline.plugins.retrievers import QA_Client
 from .models.llama_model import LlamaModel
 from .models.mpt_model import MptModel
 from .models.chatglm_model import ChatGlmModel
