+++ neural_chat/models/base_model.py
@@ -23,7 +23,6 @@ from neural_chat.pipeline.inference.inference import load_model, predict, predic
 from neural_chat.config import GenerationConfig
 from neural_chat.plugins import is_plugin_enabled, get_plugin_instance, get_registered_plugins, get_plugin_arguments
 from neural_chat.utils.common import is_audio_file
-from neural_chat.pipeline.plugins.prompts.prompt import generate_qa_prompt, generate_prompt
 
 
 def construct_parameters(query, model_name, device, config):
@@ -44,23 +43,6 @@ def construct_parameters(query, model_name, device, config):
     params["use_cache"] = config.use_cache
     params["device"] = device
     return params
-
-
-def construct_prompt(query, retriever, retrieval_type):
-    if retrieval_type == "dense":
-        documents = retriever.get_relevant_documents(query)
-        context = ""
-        for doc in documents: context = context + doc.page_content + " "
-    else:
-        documents = retriever.retrieve(query)
-        context = ""
-        for doc in documents: context = context + doc.content + " "
-    context = context.strip()
-    
-    if context != "":
-        return generate_qa_prompt(query, context)
-    else:
-        return generate_prompt(query)
 
 
 class BaseModel(ABC):
@@ -139,7 +121,7 @@ class BaseModel(ABC):
             config = GenerationConfig()
         return predict_stream(**construct_parameters(query, self.model_name, self.device, config))
 
-    def predict(self, query, config=None):
+    def predict(self, query, config=None, input_path=None):
         """
         Predict using a non-streaming approach.
 
@@ -162,20 +144,25 @@ class BaseModel(ABC):
                     if hasattr(plugin_instance, 'pre_llm_inference_actions'):
                         if plugin_name == "asr" and not is_audio_file(query):
                             continue
-                        if plugin_name == "intent_detection":
+                        if plugin_name == "safety_checker":
+                            result = plugin_instance.pre_llm_inference_actions(query)
+                            if result:
+                                return "Your query contains sensitive words, please try another query."
+                        if plugin_name == "qa_client":
                             response = plugin_instance.pre_llm_inference_actions(query,
-                                MODELS[self.model_name]["model"], MODELS[self.model_name]["tokenizer"])
+                                    MODELS[self.model_name]["model"], MODELS[self.model_name]["tokenizer"], input_path)
                         else:
                             response = plugin_instance.pre_llm_inference_actions(query)
-                        if plugin_name == "safety_checker" and response:
-                            return "Your query contains sensitive words, please try another query."
-                        elif plugin_name == "intent_detection":
-                            if 'qa' not in response.lower():
-                                query = generate_prompt(query)
-                            else:
-                                query = generate_qa_prompt(query)
-                        else:
-                            query = response
+        query = response
+                        # if plugin_name == "safety_checker" and response:
+                        #     return "Your query contains sensitive words, please try another query."
+                        # elif plugin_name == "intent_detection":
+                        #     if 'qa' not in response.lower():
+                        #         query = generate_prompt(query)
+                        #     else:
+                        #         query = generate_qa_prompt(query)
+                        # else:
+                        #     query = response
         assert query is not None, "Query cannot be None."
 
         # LLM inference
