+++ neural_chat/pipeline/plugins/intent_detector/intent_detection.py
@@ -21,23 +21,19 @@ import torch
 from ..prompts.prompt import generate_intent_prompt
 from neural_chat.plugins import register_plugin
 
-@register_plugin("intent_detection")
 class IntentDetector:
     def __init__(self):
         self.model = None
         self.tokenizer = None
 
-    def intent_detection(self, query):
+    def predict_intent(self, query, model, tokenizer):
         """Using the LLM to detect the intent of the user query."""
+        self.model=model
+        self.tokenizer= tokenizer
         prompt = generate_intent_prompt(query)
-        input_ids = self.tokenizer(prompt, return_tensors="pt").input_ids
-        input_ids = input_ids.to(self.model.device)
+        input_ids = self.tokenizer(prompt, return_tensors="pt").to(self.model.device)
         generate_ids = self.model.generate(input_ids, max_new_tokens=5, top_k=1, temperature=0.001)
         intent = self.tokenizer.batch_decode(generate_ids[:, input_ids.shape[1]:], skip_special_tokens=False,
                                         clean_up_tokenization_spaces=False)[0]
         return intent
 
-    def pre_llm_inference_actions(self, query, model, tokenizer):
-        self.model = model
-        self.tokenizer = tokenizer
-        return self.intent_detection(query)
