+++ neural_chat/pipeline/plugins/prompts/prompt.py
@@ -1,7 +1,5 @@
 """The function for generating the target prompt."""
 
-from haystack.nodes.prompt.prompt_node import PromptTemplate
-
 def generate_qa_prompt(query, context=None, history=None):
     if context and history:
         prompt = """Have a conversation with a human, answer the following questions as best you can. 
@@ -23,8 +21,10 @@ def generate_prompt(query, history=None):
     else:
         prompt = """Have a conversation with a human. You are required to generate suitable response to the user input.
                         \n\n### Input: {}\n\n### Response:""".format(query)
+    return prompt
 
 
 def generate_intent_prompt(query):
     prompt = """Please identify the intent of the provided context. 
         You may only respond with "chitchat" or "QA" without explanations or engaging in conversation.\nContext:{}\nIntent:""".format(query)
+    return prompt
