--- neural_chat/pipeline/plugins/retrievers/ranker/__init__.py
@@ -1,6 +1,0 @@
-from fastrag.rankers import BaseRanker, ColBERTRanker
-from haystack.nodes import SentenceTransformersRanker
-
-
-
-
